r121016 is the WebKit revision used by the latest stable Chrome 20.0.1132.43 / r143823
Automatic merge went fine. Build fixes follows

+++ Source/WebCore/Modules/indexeddb/IDBFactoryBackendImpl.cpp
@@ -129,8 +129,12 @@ void IDBFactoryBackendImpl::deleteDatabase(const String& name, PassRefPtr<IDBCal
     }
 
     RefPtr<IDBDatabaseBackendImpl> databaseBackend = IDBDatabaseBackendImpl::create(name, backingStore.get(), m_transactionCoordinator.get(), this, uniqueIdentifier);
-    m_databaseBackendMap.set(uniqueIdentifier, databaseBackend.get());
-    databaseBackend->deleteDatabase(callbacks);
+    if (databaseBackend) {
+        m_databaseBackendMap.set(uniqueIdentifier, databaseBackend.get());
+        databaseBackend->deleteDatabase(callbacks);
+        m_databaseBackendMap.remove(uniqueIdentifier);
+    } else
+        callbacks->onError(IDBDatabaseError::create(IDBDatabaseException::UNKNOWN_ERR, "Internal error."));
 }
 
 PassRefPtr<IDBBackingStore> IDBFactoryBackendImpl::openBackingStore(PassRefPtr<SecurityOrigin> securityOrigin, const String& dataDirectory)
@@ -177,8 +181,11 @@ void IDBFactoryBackendImpl::openInternal(const String& name, IDBCallbacks* callb
     }
 
     RefPtr<IDBDatabaseBackendImpl> databaseBackend = IDBDatabaseBackendImpl::create(name, backingStore.get(), m_transactionCoordinator.get(), this, uniqueIdentifier);
-    callbacks->onSuccess(RefPtr<IDBDatabaseBackendInterface>(databaseBackend.get()).release());
-    m_databaseBackendMap.set(uniqueIdentifier, databaseBackend.get());
+    if (databaseBackend) {
+        m_databaseBackendMap.set(uniqueIdentifier, databaseBackend.get());
+        databaseBackend->openConnection(callbacks);
+    } else
+        callbacks->onError(IDBDatabaseError::create(IDBDatabaseException::UNKNOWN_ERR, "Internal error."));
 }
 
 } // namespace WebCore
